fix(feed): show the Ignore menu only in the thread view
The post 'more' menu (Ignore post / Ignore person) was wired on every feed
row but absent from the thread/detail view. Move it: ignoring is a considered
action, so it now lives only where you've tapped into a post. Extract the menu
into a shared usePostActions hook used by ThreadView, and remove it (plus the
now-dead openPostActions / removeIgnoredItems / feedItemMatchesIgnoreTarget
helpers and ignore-action selectors) from HomeFeed. The feed still hides
ignored posts/people reactively via the ignore store, so no per-row removal is
needed.
